requirements.txt
# Core dependencies
torch>=2.2.0
torchaudio>=2.2.0
numpy>=1.24.0
pandas>=2.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
moviepy>=1.0.3

# NLP
transformers>=4.30.0
scikit-learn>=1.3.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
pyyaml>=6.0

# Optional: for visualization
tensorboard>=2.13.0
